lab-01-fundamentals/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0

